refactor: 영상 카드 이모지 버튼 이벤트를 템플릿에서 등록하도록 변경
