+++ src/js/ui/Main.js
@@ -1,19 +1,19 @@
 import { store } from '../domain/store';
-import { $, $$ } from '../utils/dom';
-import { showExceptionSnackBar } from '../utils/snackBar';
+import { $ } from '../utils/dom';
 import { MESSAGE, STATE, STORAGE_KEY } from '../constants';
 import { getVideoTemplate } from './template';
 
 export default class Main {
   constructor() {
     this.tab = STATE.WILL;
+    this.$videoContainer = $('.video-container');
     this.renderVideo(this.tab);
     this.addNavButtonClickEvent();
   }
 
   renderNoSavedVideo() {
-    $('.video-container').replaceChildren();
-    $('.video-container').insertAdjacentHTML(
+    this.$videoContainer.replaceChildren();
+    this.$videoContainer.insertAdjacentHTML(
       'beforeend',
       MESSAGE.NO_SAVED_VIDEO,
     );
@@ -29,17 +29,14 @@ export default class Main {
       return;
     }
 
-    const $videoContainer = $('.video-container');
-    $videoContainer.replaceChildren();
+    this.$videoContainer.replaceChildren();
     const $saveVideoList = document.createElement('ul');
     $saveVideoList.classList.add('save-video-list');
     $saveVideoList.insertAdjacentHTML(
       'beforeend',
       getVideoTemplate(filteredVideos),
     );
-    $videoContainer.appendChild($saveVideoList);
-    this.addWatchButtonClickEvent();
-    this.addDeleteButtonClickEvent();
+    this.$videoContainer.appendChild($saveVideoList);
   }
 
   addNavButtonClickEvent() {
@@ -60,47 +57,5 @@ export default class Main {
       this.tab = STATE.WATCHED;
       this.renderVideo(this.tab);
     });
-  }
-
-  addWatchButtonClickEvent() {
-    $$('.watch-button').forEach(button =>
-      button.addEventListener('click', e => {
-        const { videoId } = e.target.dataset;
-        const saveVideos = store.getLocalStorage(STORAGE_KEY);
-
-        store.removeLocalStorage(STORAGE_KEY);
-        saveVideos.forEach(video => {
-          if (video.videoId === videoId) {
-            video.watched = !video.watched;
-          }
-          store.setLocalStorage(STORAGE_KEY, video);
-        });
-
-        showExceptionSnackBar(MESSAGE.MODIFY_COMPLETE);
-
-        this.renderVideo(this.tab);
-      }),
-    );
-  }
-
-  addDeleteButtonClickEvent() {
-    $$('.delete-button').forEach(button =>
-      button.addEventListener('click', e => {
-        if (!confirm(MESSAGE.ASK_DELETE)) return;
-        const { videoId } = e.target.dataset;
-        const saveVideos = store.getLocalStorage(STORAGE_KEY);
-
-        store.removeLocalStorage(STORAGE_KEY);
-        saveVideos.forEach(video => {
-          if (video.videoId !== videoId) {
-            store.setLocalStorage(STORAGE_KEY, video);
-          }
-        });
-
-        showExceptionSnackBar(MESSAGE.MODIFY_COMPLETE);
-
-        this.renderVideo(this.tab);
-      }),
-    );
   }
 }
